Fix non-deterministic results
FlowDroid does remove abstractions from the results if they are
entailed by another abstraction to prevent building unnecessary
paths. The entails check did not correctly take the whole domain
into consideration, such that abstractions with the cutAP flag
were considered entailed by flows without the cutAP flag. This
lead to removal of non-isomorphic abstraction graphs and if
these two abstraction graphs had different sources in it, also
to non-deterministic results, depending on the iteration order
of the result set.

This commit fixes this issue. The non-determinism described in
the paper "TaintBench: Automatic real-world malware benchmarking
of Android taint analyses" seems to be related to the same issue
and should be fixed with this as well.